main: Product scrape POC

+++ python/src/service/synchronizer/model_synchorizer.py
@@ -1,3 +1,5 @@
+from src.service.dto.category_products import CategoryProducts
+from src.service.dto.product import Product
 from src.service.util.category_util import CategoryUtil
 from src.repository.category_repository import category_repository
 from src.repository.model_repository import model_repository
@@ -5,28 +7,48 @@ from src.model.brand import Brand
 from src.model.model import Model
 from src.database import db
 
-
 class ModelSynchronizer:
     def __init__(self):
         self.database_session = db.get_session()
 
-    def sync(self) -> None:
-        existing_categories = category_repository.fetch_all()
-        existing_models = model_repository.fetch_all()
+    def sync_from_products(self, category_products: list[CategoryProducts]) -> None:
         models_to_keep = []
-        for category in existing_categories:
-            model_name = CategoryUtil.extract_model_name(category.name)
+        all_new_names = []
+        for category_product in category_products: #type: CategoryProducts
+            new_model_names = []
+            existing_models = model_repository.fetch_all_by_category_id(category_product.category_id)
+            for product in category_product.products: #type: Product
+                model_name = product.model
+                existing_model = self.get_existing_model(model_name, existing_models)
+                if existing_model is not None:
+                    models_to_keep.append(existing_model.id)
 
-            existing_model = self.get_existing_model(model_name, existing_models)
-            if existing_model is not None:
-                models_to_keep.append(existing_model.id)
+                    continue
 
-                continue
-            model = Model(name=model_name, brand_id=category.brand_id)
-            self.database_session.add(model)
+                new_model_names.append(model_name)
 
-        db.deleteWithCondition(Model, ~Model.id.in_(models_to_keep))
+            unique_new_names = self.get_unique_names(new_model_names, all_new_names)
+            print("Unique model names: ", unique_new_names, " category_id: ", category_product.category_id)
+            # exit()
+            for model_name in unique_new_names:
+                model = Model(name = model_name, category_id = category_product.category_id)
+                self.database_session.add(model)
+
+            all_new_names = all_new_names + list(unique_new_names)
+
         self.database_session.commit()
+
+        if models_to_keep:
+            db.deleteWithCondition(Model, ~Model.id.in_(models_to_keep))
+
+    def get_unique_names(self, new_names: list[str], all_names: list[str])-> set[str]:
+        unique_names = []
+        for new_name in new_names:
+            if new_name in all_names:
+                continue
+            unique_names.append(new_name)
+
+        return set(unique_names)
 
     def get_existing_model(self, model_name: str, existing_models: list[Model]) -> Model | None:
         for existing_model in existing_models:
